+++ src/server/public.ts
@@ -1,2 +1,13 @@
-export type { AuthConfig, DatabaseConfig, DeepJsonServerConfig, FilesConfig, GraphqlConfig, MemoryFile, OpenapiConfig, PackageConfig, ServerConfig, Storage } from './config.js';
+export type {
+  AuthConfig,
+  DatabaseConfig,
+  DeepJsonServerConfig,
+  FilesConfig,
+  GraphqlConfig,
+  MemoryFile,
+  OpenapiConfig,
+  PackageConfig,
+  ServerConfig,
+  Storage,
+} from './config.js';
 export { createServer, type ServerFacade } from './create.js';